+++ Ganga/client/src/Components/Panels/AdminPanel/vendorList.jsx
@@ -4,6 +4,7 @@ import { DataGrid } from "@mui/x-data-grid";
 import { Link} from "react-router-dom";
 import { TiDeleteOutline } from "react-icons/ti";
 import { BsPencilSquare } from "react-icons/bs";
+import { Swal } from 'sweetalert2';
 
 import RevenueChart from "./aCharts/Revenue";
 import s from "./admin.module.css";
@@ -42,8 +43,27 @@ export default function VendorList({ vendors, products}) {
   ];
 
   function handleDelete(id) {
-    dispatch(deleteUser(id));
-    setRows(rows.filter((i) => i.id !== id));
+    Swal.fire({
+      title: 'Estas seguro?',
+      text: "Se borraran todos los datos del vendedor.",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      cancelButtonText: 'Cancelar',
+      confirmButtonText: 'Confirmar'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire(
+          'Eliminado!',
+          'El vendedor ha sido eliminado.',
+          'success'
+          )
+          dispatch(deleteUser(id));
+          setRows(rows.filter((i) => i.id !== id));
+      }
+    })
+    window.location.reload();
   }
 
   const Rows = vendors?.map((u) => {
